Restrict swagger-ui rendering to its own instance
This modification allows multiple instances of swagger-ui on the same
page to work properly.
By default, swagger-ui always renders the API resources in the first
 #resources it finds in the DOM. This modification adds some context to
the render. It will look for a DOM element with the ID passed as a
parameter during the swagger-ui initialization, and then look for the
 #resources in its children nodes.